cleanup loader callback
